+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Attempts to stratify on a `Surv` object now error more informatively (#230). 
 
+* Exposed `pool` argument from `make_strata()` in user-facing resampling functions (#229).
+
 
 # rsample 0.0.9